fix: renaming some files, removed some css and resolved comments
