+++ Phonebook.Frontend/src/app/shared/components/user/user-detail/user-detail.component.scss
@@ -18,16 +18,6 @@ mat-card-header {
     display: flex;
   }
 }
-
-.blur {
-  background-size: cover;
-  -webkit-filter: blur(4px);
-  -moz-filter: blur(4px);
-  -ms-filter: blur(4px);
-  -o-filter: blur(4px);
-  filter: blur(4px) grayscale(0.01);
-}
-
 mat-card-content {
   .pb-roomplan {
     width: 100%;
@@ -59,7 +49,6 @@ mat-card-content {
     align-content: center;
     flex-flow: column;
   }
-
 
   mat-list {
     flex: 1;
