+++ Phonebook.Frontend/src/app/shared/dialogs/user-information/Incorrect-user-information.component.html
@@ -1,4 +1,4 @@
-<div class="pb-text-align-center>
+<div class="pb-text-align-center">
   <div class="center">
     <h2 mat-dialog-title i18n="User-InformationDialog|Headline@@Headline">User Information</h2>
     <div mat-dialog-content *ngIf="onMyProfile">
