--- Phonebook.Frontend/src/app/shared/dialogs/user-information/user-information.component.scss
@@ -1,14 +1,0 @@
-.center {
-    text-align: center;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .actions {
-    text-align: center;
-    top: 5px;
-  }
-  .actions > button {
-    width: 100%;
-  }
